engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.4.2
pandas==2.2.1
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.6.4
httpx==0.26.0
python-dotenv==1.0.1
aioredis==2.0.1
